refactor: port over profile_data_asset
